Update index.tsx
I shared this link with someone and the preview displayed the note "description will go into a meta tag in". I wrote a comment to remember to make that change in the future but also changed the current description so that it looks sharp for anyone that sees a preview before that happens.